Extract x86-specific trap handler code
